# Make line break tests run faster