Currently, we load line break data *every time* we call `is_break` or
`get_linebreak_property`. It is sufficient to load the data once per test.

On my machine, the running time of `cargo test break_rule` is reduce from 2.66s
to 0.05s, and `cargo test linebreak_property` from 0.62s to 0.05s.